ImageToTextQwen/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyMuPDF
